Feature: Create new controller for user to login or register and the user animal controller have own controller
