+++ API/Auth/AuthenticationController.cs
@@ -1,0 +1,74 @@
+﻿using Application.Commands.Users.Register;
+using Application.Dtos.Users;
+using Application.Dtos.Validation;
+using Application.Exceptions.Authorize;
+using Application.Queries.Users.Login;
+using Application.Validators.User;
+using MediatR;
+using Microsoft.AspNetCore.Mvc;
+
+namespace API.Controllers.AuthController
+{
+    [Route("api/[controller]")]
+    [ApiController]
+    public class AuthenticationController : ControllerBase
+    {
+        internal readonly IMediator _mediator;
+        internal readonly UserValidator _userValidator;
+
+        public AuthenticationController(IMediator mediator,
+                               UserValidator userValidator)
+        {
+            _mediator = mediator;
+            _userValidator = userValidator;
+        }
+
+        [HttpPost("register")]
+        [ProducesResponseType(typeof(UserDto), StatusCodes.Status200OK)]
+        [ProducesResponseType(typeof(Errors), StatusCodes.Status400BadRequest)]
+        public async Task<IActionResult> Register([FromBody] UserDto userToRegister)
+        {
+            var inputValidation = _userValidator.Validate(userToRegister);
+
+            if (!inputValidation.IsValid)
+            {
+                return BadRequest(inputValidation.Errors.ConvertAll(errors => errors.ErrorMessage));
+            }
+
+            try
+            {
+                return Ok(await _mediator.Send(new RegisterUserCommand(userToRegister)));
+            }
+            catch (ArgumentException e)
+            {
+                return BadRequest(e.Message);
+            }
+        }
+
+        [HttpPost("login")]
+        [ProducesResponseType(typeof(TokenDto), StatusCodes.Status200OK)]
+        [ProducesResponseType(typeof(Errors), StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(StatusCodes.Status401Unauthorized)]
+        public async Task<IActionResult> Login([FromBody] UserDto userToLogin)
+        {
+            var inputValidation = _userValidator.Validate(userToLogin);
+
+            if (!inputValidation.IsValid)
+            {
+                return BadRequest(inputValidation.Errors.ConvertAll(errors => errors.ErrorMessage));
+            }
+
+            try
+            {
+                string token = await _mediator.Send(new LoginUserQuery(userToLogin));
+
+                return Ok(new TokenDto { TokenValue = token });
+            }
+            catch (UnAuthorizedException ex)
+            {
+                return BadRequest(ex.Message);
+            }
+        }
+
+    }
+}
